Separate fork runtime part mapping
The fork runtime stream module mixed child-run orchestration with AG-UI event-to-fork-part mapping. This extracts the mapper and recovered-part state into a dedicated module while re-exporting the old symbols from fork-runtime-stream so existing imports keep working.

Constraint: Preserve the public veryfront/agent export surface.

Rejected: Split all fork runtime step execution in this PR | The mapper boundary is independently testable and keeps the diff reviewable.

Confidence: high

Scope-risk: narrow

Directive: Keep event-to-part mapping behavior in fork-runtime-part-mapper; keep step orchestration in fork-runtime-stream unless a follow-up extracts a separate step runner.

Tested: deno test --no-check --allow-all src/agent/streaming/fork-runtime-part-mapper.test.ts src/agent/streaming/fork-runtime-stream.test.ts

Tested: deno check src/agent/streaming/fork-runtime-part-mapper.ts src/agent/streaming/fork-runtime-part-mapper.test.ts src/agent/streaming/fork-runtime-stream.ts src/agent/streaming/fork-runtime-stream.test.ts src/utils/version-constant.ts

Tested: git diff --check

Tested: deno task verify:quick

Not-tested: Hosted CI pending PR.
